server/requirements-frozen.txt
click==8.0.1
cycler==0.10.0
Cython==0.29.24
exif==1.3.2
Flask==2.0.1
gunicorn==20.1.0
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
loguru==0.5.1
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pycocotools==2.0.2
pydantic==1.5.1
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
sk-video==1.1.10
Werkzeug==2.0.1
../packages/openvino_person_detection_retail-0.2.1+cpu-py3-none-any.whl
